Поделил ответсвенность тестов - Некоторые тесты перекочевали в InWord аналоги
