+++ cs/MarkdownTests/Parser/Rules/TagRules/BoldRuleTest.cs
@@ -54,20 +54,6 @@ public class BoldRuleTest
         node.NodeType.Should().Be(NodeType.BOLD);
     }
     
-    [TestCase("Bold tag in wo__rd__ end", 7, ExpectedResult = "rd")]
-    [TestCase("Bold tag in __wo__rd begin", 6, ExpectedResult = "wo")]
-    [TestCase("Bold tag in w__or__d center", 7, ExpectedResult = "or")]
-    public string BoldRule_Match_ShouldMatchTagInWord(string text, int begin)
-    {
-        var tokens = tokenizer.Tokenize(text);
-        
-        var node = rule.Match(tokens, begin) as TagNode;
-        
-        node.Should().NotBeNull();
-        node.Children.Should().ContainSingle(n => n.NodeType == NodeType.TEXT);
-        return node.Children.First().As<TextNode>().Text;
-    }
-    
     [TestCase("Bold tag with__123__numbers", 6)]
     [TestCase("Numbers that separates with underscores 12__34__56", 10)]
     public void BoldRule_Match_ShouldNotMatchTextWithNumbers(string text, int begin)
